Add `config.rs`
